requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
pikepdf==9.7.0
pdfplumber==0.11.6
watchdog==6.0.0
python-multipart==0.0.20
aiofiles==24.1.0
platformdirs==4.3.6
cryptography==44.0.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.161.0
